+++ modules/base.nix
@@ -33,7 +33,7 @@ let
     };
   };
 in {
-  imports = [ "cachix" "flakes" "impermanence" "nix-index" "termite" "vm" "vscode" "xdg" ];
+  imports = [ "alacritty" "cachix" "flakes" "impermanence" "nix-index" "termite" "vm" "vscode" "xdg" ];
 
   nixosModule = { config, configRevision, user, pkgs, ... }: {
     imports = [ shared ];