+++ modules/personal.nix
@@ -1,9 +1,16 @@
 {
   imports = [ "i18n" "ios" "pim" ];
 
-  nixosModule = { pkgs, ...}: {
+  nixosModule = { user, pkgs, ...}: {
     services.mullvad-vpn.enable = true;
     services.mullvad-vpn.package = pkgs.mullvad-vpn;
+
+    services.resilio.enable = true;
+    services.resilio.listeningPort = 44444;
+    services.resilio.enableWebUI = true;
+    services.resilio.httpListenAddr = "0.0.0.0";
+
+    users.users.${user}.extraGroups = [ "rslsync" ];
   };
 
   hmModule = { config, pkgs, ... }: {